Added group and user service.

+++ application/src/main/java/com/etsmtl/codecrusade/repository/ClassGroupRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClassGroupRepository extends PagingAndSortingRepository<ClassGroup, Long> {
+public interface ClassGroupRepository extends PagingAndSortingRepository<ClassGroup, Integer> {
 }
